Puppet apply script complains if you have directories that are world writable. So mount it dmode=775.
glibc-devel is required to complete the Oracle database install. This is not present in bento/oracle-7.7 so install it manually at provision time.

Strange behaviour with vbguest when attempting to chmod +x on a file that is already executable. It removes the executable flag! Adding a check will prevent this issue.