Resolve #92 (#93)
* Fix default config.rb.example to use a real directory

* Add empty download dir as expected by config.rb

* Add empty download dir as expected by config.rb

* Added comments to remind users to make the change to VP1 for IH and FSCM

* More instances of dpk vs. dpks. Fixes #66

* Resolve #72

* Fix for #82

* Resolve vbguest issue

* Removed a rogue blank line

* Puppet apply script complains if you have directories that are world writable. So mount it dmode=775.

glibc-devel is required to complete the Oracle database install. This is not present in bento/oracle-7.7 so install it manually at provision time.

Strange behaviour with vbguest when attempting to chmod +x on a file that is already executable. It removes the executable flag! Adding a check will prevent this issue.

* Forgot to include glibc-devel. It is there now

* 8.58 has a new requirement for the response.cfg file.

Add it here.

Resolves #92

Co-authored-by: Dan Iverson <dan@daniverson.com>